WPE-mesa: Fix Wayland client flushing
Fixes a bug where the Wayland client display was flushed before
dispatching any pending messages. This is incorrect because the
dispatching of the messages might result in additional outbound messages
being queued up, and these messages also need to be flushed before the
thread starts blocking on poll().

In addition, the poll event check in the EGL rendering backend was
incorrectly treating G_IO_IN as an error, when it should have been using
G_IO_ERR.